pyproject.toml
[project]
name = "financial-intelligence-agent"
version = "0.1.0"
description = "Financial Intelligence Agentic System with Text2SQL, LangGraph orchestration, and MCP integration"
authors = [{ name = "Financial Analytics Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.1.0",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.30",
    "snowflake-connector-python>=3.10.0",
    "boto3>=1.34.0",
    "pandas>=2.2.0",
    "tenacity>=8.2.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "orjson>=3.10.0",
    "cryptography>=42.0.0",
    "pyyaml>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "types-redis>=4.6.0",
    "types-requests>=2.32.0"
]

[project.scripts]
fia-agent = "fia_agent.main:run"

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W", "UP", "N", "ASYNC", "DTZ"]
ignore = ["E203", "E266", "E501"]
